+++ torch/ao/sparsity/_experimental/pruner/images/prune_1.png
