+++ torch/ao/sparsity/_experimental/pruner/images/prune_2.png
